Rename liskService action/reducer as service
