+++ src/store/reducers/index.js
@@ -2,4 +2,4 @@ export { default as accounts } from './accounts';
 export { default as transactions } from './transactions';
 export { default as loading } from './loading';
 export { default as settings } from './settings';
-export { default as liskService } from './liskService';
+export { default as service } from './service';
